requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0.0
urllib3>=2.0.0
selenium>=4.15.0
fake-useragent>=1.4.0
price-parser>=0.3.4
validators>=0.22.0
tenacity>=8.2.0

# Optional: for REST API (api_example.py)
# Uncomment if you want to use the API wrapper
# flask>=3.0.0
# flask-cors>=4.0.0

